Fix recursive ocfl tar task creation
Filter out anything that is not a file on the root of the inbox, for any
inbox. This caused some issues with ocfl tar tasks being generated and a
huge tree of rejected/rejected/rejected etc.

Also verified that an inbox item is actually known in the database,
otherwise rejecting it.